Extract temperatur.nu plugin into a separate file
